+++ Event/ControllerArgumentsEvent.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\HttpKernel\Event;
 
+use Symfony\Component\ExpressionLanguage\Expression;
+use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 
@@ -115,5 +117,14 @@ final class ControllerArgumentsEvent extends KernelEvent
     public function getAttributes(?string $className = null): array
     {
         return $this->controllerEvent->getAttributes($className);
+    }
+
+    public function evaluate(mixed $value, ?ExpressionLanguage $expressionLanguage): mixed
+    {
+        if (!$value instanceof \Closure && !$value instanceof Expression) {
+            return $value;
+        }
+
+        return $this->controllerEvent->evaluate($value, $expressionLanguage, $this->getNamedArguments());
     }
 }